requirements.txt
pandas>=2.0.0
numpy>=1.24.0
zeyrek>=0.1.3
snowballstemmer>=2.2.0
scikit-learn>=1.3.0
regex>=2023.0
tqdm>=4.65.0
turkish-deasciifier @ git+https://github.com/emres/turkish-deasciifier.git
